refactored contour data by creating a ProcessedContourData interface
